fix: resolve React Hook useEffect dependency warning in Navbar
- Move navItems constant outside component scope
- Fixes react-hooks/exhaustive-deps warning
- Improves performance by avoiding array recreation on each render
- navItems is static configuration data and doesn't need component scope
