+++ app/components/ui/globals.css
@@ -136,6 +136,10 @@ body {
   * {
     @apply border-border outline-ring/50;
   }
+  html {
+  scroll-behavior: smooth;
+}
+
   body {
     @apply bg-background text-foreground;
   }
